test(ext-openai): migrate OpenAI runtime tests from core runtime-loader
Move all OpenAI-specific test blocks out of src/provider/runtime-loader.test.ts
into extensions/ext-openai/src/openai-provider.test.ts, referencing the (not-yet-
existing) createOpenAI* exports on openai-provider.ts (added in Task 14).
